:recycle: [refactor]: Update AuthContext Component
- 기존 [로그인 확인 API] 를 custom hook 으로 모듈화 시킴
-> useAuthFetch.ts
